pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Yoskoslowich Fernandes <yoskos@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.112.1"}
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.pytest.ini_options]
pythonpath = "."
#addopts = "-p no:warnings"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.taskipy.tasks]
run = "fastapi dev fast_zero/app.py"
lint = "ruff check . && ruff check . --diff"
#format = "ruff check . --fix && ruff format ."
format = "ruff format ."
pre_test = "task format"
test = "pytest --cov=fast_zero -vv"
post_test = "coverage html"
migrate = "alembic upgrade head"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
